+++ torch/_inductor/select_algorithm.py
@@ -1351,12 +1351,8 @@ class TritonTemplate(KernelTemplate):
         super().__init__(name)
         self.grid = grid
         self.template = self._template_from_string(source)
-        try:
-            assert name not in self.all_templates
-        except:
-            breakpoint()
         assert name not in self.all_templates, "duplicate template name"
-        print(f"Registering template {name}")
+        log.debug("Registering template {}", name)
         TritonTemplate.all_templates[name] = self
         self.debug = debug
         self._cache_codegen_enabled_for_template = cache_codegen_enabled_for_template
